machines: refactoring; move vm and dummy vm out of hostvmslist
