+++ pkg/ovirt/components/ClusterVms.jsx
@@ -23,7 +23,7 @@ import cockpit from 'cockpit';
 import CONFIG from '../config.es6';
 
 import { Listing, ListingRow } from "cockpit-components-listing.jsx";
-import { StateIcon } from "../../machines/hostvmslist.jsx";
+import StateIcon from "../../machines/components/vm/stateIcon.jsx";
 import DropdownButtons from "../../machines/components/dropdownButtons.jsx";
 
 import { toGigaBytes, valueOrDefault, isSameHostAddress } from '../helpers.es6';
